+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Show unresolved aliases on "Aliases" page.
+- Add button for removing declared aliases.
 
 ### Changed